add events to `Account` and `Client`
